+++ lib/kitchen/directions/bake_numbered_table/v2.rb
@@ -26,6 +26,8 @@ module Kitchen::Directions::BakeNumberedTable
           </div>
         HTML
       )
+
+      table.parent.add_class('os-timeline-table-container') if table.has_class?('timeline-table')
     end
   end
 end