+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Small fixes to return when no elements are found and not add an empty wrapper in `BakeChapterReferences` and
 `BakeFreeResponse` (minor)
 * Adding class `os-timeline-table-container` to numbered tables when required (minor)
+* Fix `BakeExample` to catch the multiple solutions to one exercise (patch)
 
 ## [11.1.0] - 2021-08-30